+++ packages/rest-api/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.47](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.46...@synapsecns/rest-api@1.0.47) (2024-01-19)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.0.46](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.45...@synapsecns/rest-api@1.0.46) (2024-01-16)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
 ## [1.0.45](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.44...@synapsecns/rest-api@1.0.45) (2024-01-12)
 
 **Note:** Version bump only for package @synapsecns/rest-api