+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.20](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.19...@synapsecns/sdk-router@0.3.20) (2024-01-19)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.19](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.18...@synapsecns/sdk-router@0.3.19) (2024-01-16)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.3.18](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.17...@synapsecns/sdk-router@0.3.18) (2024-01-12)
 
 **Note:** Version bump only for package @synapsecns/sdk-router